# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlacementPulse API - Future-ready placement preparation platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SkillTrendProperty = typeof SkillTrendProperty[keyof typeof SkillTrendProperty];


export const SkillTrendProperty = {
  rising: 'rising',
  stable: 'stable',
  declining: 'declining',
} as const;

export interface Skill {
  id: number;
  name: string;
  category: string;
  /** Score 0-100 representing current hiring demand */
  currentDemandScore: number;
  /** Predicted demand score 2 years from now */
  futureDemandScore: number;
  trend: SkillTrendProperty;
  description: string;
  /**
     * JSON array of resource links
     * @nullable
     */
  resources?: string | null;
  /** @nullable */
  avgSalary?: number | null;
  /** @nullable */
  jobCount?: number | null;
  createdAt?: string;
}

export type SkillInputTrend = typeof SkillInputTrend[keyof typeof SkillInputTrend];


export const SkillInputTrend = {
  rising: 'rising',
  stable: 'stable',
  declining: 'declining',
} as const;

export interface SkillInput {
  name: string;
  category: string;
  currentDemandScore: number;
  futureDemandScore: number;
  trend: SkillInputTrend;
  description: string;
  resources?: string;
  avgSalary?: number;
  jobCount?: number;
}

export interface Role {
  id: number;
  title: string;
  domain: string;
  avgSalary: number;
  description: string;
  /** @nullable */
  openings?: number | null;
}

export type RoleSkillImportance = typeof RoleSkillImportance[keyof typeof RoleSkillImportance];


export const RoleSkillImportance = {
  must_have: 'must_have',
  good_to_have: 'good_to_have',
  bonus: 'bonus',
} as const;

export interface RoleSkill {
  id: number;
  roleId: number;
  skillId: number;
  importance: RoleSkillImportance;
  skillName: string;
  skillCategory?: string;
}

export interface Profile {
  id: number;
  name: string;
  email: string;
  targetRoleId: number;
  /** @nullable */
  targetRoleTitle?: string | null;
  graduationYear: number;
  college: string;
  createdAt?: string;
}

export interface ProfileInput {
  name: string;
  email: string;
  targetRoleId: number;
  graduationYear: number;
  college: string;
}

export interface ProfileUpdate {
  name?: string;
  targetRoleId?: number;
  graduationYear?: number;
  college?: string;
}

export type ProfileSkillProficiencyLevel = typeof ProfileSkillProficiencyLevel[keyof typeof ProfileSkillProficiencyLevel];


export const ProfileSkillProficiencyLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface ProfileSkill {
  id: number;
  profileId: number;
  skillId: number;
  proficiencyLevel: ProfileSkillProficiencyLevel;
  skillName: string;
  skillCategory: string;
  /** @nullable */
  currentDemandScore?: number | null;
  /** @nullable */
  futureDemandScore?: number | null;
  /** @nullable */
  trend?: string | null;
}

export type ProfileSkillInputProficiencyLevel = typeof ProfileSkillInputProficiencyLevel[keyof typeof ProfileSkillInputProficiencyLevel];


export const ProfileSkillInputProficiencyLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface ProfileSkillInput {
  skillId: number;
  proficiencyLevel: ProfileSkillInputProficiencyLevel;
}

export type RoadmapStepPriority = typeof RoadmapStepPriority[keyof typeof RoadmapStepPriority];


export const RoadmapStepPriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface RoadmapStep {
  skillId: number;
  skillName: string;
  priority: RoadmapStepPriority;
  timeEstimateWeeks: number;
  reason: string;
  currentDemandScore?: number;
  futureDemandScore?: number;
  trend?: string;
}

export interface Roadmap {
  profileId: number;
  targetRole: string;
  /** 0-100 score indicating placement readiness */
  readinessScore: number;
  gapSkills: RoadmapStep[];
  steps: RoadmapStep[];
}

export interface Assessment {
  id: number;
  profileId: number;
  skillId: number;
  /** 1-5 self rating */
  selfRating: number;
  /** @nullable */
  notes?: string | null;
  skillName: string;
  createdAt: string;
}

export interface AssessmentInput {
  profileId: number;
  skillId: number;
  selfRating: number;
  notes?: string;
}

export interface DashboardStats {
  totalSkills: number;
  risingSkillsCount: number;
  decliningSkillsCount: number;
  stableSkillsCount?: number;
  topCategory: string;
  /** Average increase from current to future demand across all rising skills */
  avgFutureBoost: number;
}

export interface SkillTrend {
  skillId: number;
  skillName: string;
  category: string;
  trend: string;
  currentDemandScore: number;
  futureDemandScore: number;
  growthPercent?: number;
}

export type ListSkillsParams = {
category?: string;
search?: string;
};