# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlacementPulse API - Future-ready placement preparation platform
 * OpenAPI spec version: 0.1.0
 */

export interface AssessmentInput {
  profileId: number;
  skillId: number;
  selfRating: number;
  notes?: string;
}