# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PlacementPulse API - Future-ready placement preparation platform
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  email: string;
  targetRoleId: number;
  /** @nullable */
  targetRoleTitle?: string | null;
  graduationYear: number;
  college: string;
  createdAt?: string;
}